Handle pod failures for all policies (#189)
If a pod is in phase failure we have to create a new one.
Currently it was assumed the pod would restart due to a RestartPolicy on the pod level
This doesn't work if the pod fails for a system reason.

(cherry picked from commit 8f0ddb5aa96e6d1950757355fdc048eea7d315ad)